+++ pkgs/frontend/components/Icons/AddressIcon.tsx
@@ -1,0 +1,13 @@
+const EvaluateIcon = (
+    props: JSX.IntrinsicAttributes & React.SVGProps<SVGSVGElement>
+) => (
+    <svg xmlns='http://www.w3.org/2000/svg' width='24px' height='24px' {...props}>
+    <g>
+        <path d='M9.586 15a2 2 0 0 1 1.284.467l.13.119 1 1 1-1a2 2 0 0 1 1.238-.578l.176-.008h4.024a3 3 0 0 1 2.862 2.1l.049.172.466 1.864a1.5 1.5 0 0 1-1.324 1.858L20.36 21H3.64a1.5 1.5 0 0 1-1.481-1.735l.026-.129.466-1.864a3 3 0 0 1 2.732-2.267L5.562 15h4.024ZM12 2c.784 0 1.661.19 2.38.391 1.596.445 2.542 1.896 2.615 3.415l.005.19v4.007c0 1.588-.957 3.143-2.62 3.606-.719.2-1.596.391-2.38.391-.784 0-1.661-.19-2.38-.391-1.596-.445-2.542-1.896-2.615-3.415L7 10.004V5.996c0-1.587.957-3.143 2.62-3.606C10.339 2.191 11.216 2 12 2Z'/>
+    </g>
+</svg>
+);
+
+export default EvaluateIcon;
+
+
