+++ pkgs/frontend/components/Icons/VerifyIcon.tsx
