+++ pkgs/frontend/pages/verify.tsx
@@ -1,10 +1,11 @@
+import { NavigationSidebar } from "@/components/navigation/navigationSidebar";
 
 export default function Verify() {
-    return (
-      <>
-          <h1>Verify</h1>
-          <p>This is Verify ページ</p>
-      </>
-    )
-  }
-  
+  return (
+    <div className="h-screen w-screen flex flex-row">
+      <NavigationSidebar />
+      <h1>Verify</h1>
+      <p>This is Verify page</p>
+    </div>
+  );
+}
