+++ pkgs/frontend/tailwind.config.js
@@ -18,6 +18,25 @@ module.exports = {
 
       fontSize: {
         "6xl": "56px",
+        "Title": ['1.5rem', {
+          letterSpacing: '0.028em',
+          fontWeight: '400',
+        }],
+        "Button": ['1rem', {
+          letterSpacing: '0.028em',
+          fontWeight: '600',
+        }],
+        "Input": ['1.5rem', {
+          fontWeight: '400',
+          lineHeight: '1rem',
+        }],
+        "InputLabel": ['1.5rem', {
+          fontWeight: '400',
+        }],
+        "InputLabelFocus": ['0.9rem', {
+          fontWeight: '400',
+          lineHeight: '0.625rem',
+        }],
       },
 
       colors: {
@@ -25,8 +44,16 @@ module.exports = {
         Primary20: "#463D57",
         Primary30: "#665E75",
         Primary40: "#9691A0",
+        Input10: "#144F99",
+        Input20: "#807C87",
+        Input30: "#EBEFF5",
+        "white": "#FCFDFF",
         "pink-700": "#C3ABC2",
         "blue-400": "#6AA3EB",
+      },
+
+      backgroundImage: {
+        LoginGradient: 'linear-gradient(-40deg, #6AA3EB, #C3ABC2)',
       },
 
       opacity: {
@@ -37,6 +64,10 @@ module.exports = {
 
       borderWidth: {
         3: "3px",
+      },
+
+      boxShadow: {
+        'lg': '0 16px 32px 4px rgba(54, 44, 73, 0.2)',
       },
     },
   },
